First implementation of getReport

+++ src/main/java/pcd/assignment/tasks/executors/model/data/FileInfo.java
@@ -4,9 +4,9 @@ import java.io.File;
 
 public class FileInfo {
     private final File file;
-    private final int lineCount;
+    private final long lineCount;
 
-    public FileInfo(File file, int lineCount) {
+    public FileInfo(File file, long lineCount) {
         this.file = file;
         this.lineCount = lineCount;
     }
@@ -15,7 +15,7 @@ public class FileInfo {
         return this.file;
     }
 
-    public int getLineCount() {
+    public long getLineCount() {
         return this.lineCount;
     }
 }
